requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
python-dotenv==1.0.0
web3==6.11.3
aiohttp==3.9.1
asyncio-throttle==1.0.2
scipy==1.11.4
scikit-learn==1.3.2
